pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "lacquertutor"
version = "0.1.0"
description = "Contract-centered mixed-initiative agent for lacquer craft learning"
requires-python = ">=3.11"
dependencies = [
    "openai-agents>=0.10",
    "openai>=2.0",
    "mem0ai>=2.0.0b1",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "httpx>=0.27",
    "fastapi>=0.115",
    "uvicorn>=0.30",
    "typer[all]>=0.12",
    "rich>=13.0",
    "structlog>=24.0",
    "tenacity>=8.0",
    "aiosqlite>=0.20",
]

[project.optional-dependencies]
eval = ["scipy>=1.11"]
retrieval = [
    "qdrant-client>=1.9",
]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "scipy>=1.11",
    "qdrant-client>=1.9",
]

[project.scripts]
lacquertutor = "lacquertutor.cli.app:app"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
